De-couple model interfaces from response interfaces
