+++ src/WebPageInterface.php
@@ -17,16 +17,9 @@ interface WebPageInterface extends WebResourceInterface
     public function getCharacterSet(): ?string;
 
     /**
-     * Returns the character set found in the document i.e. the body of the WebResourceInterface instance.
+     * Returns the character set found in the document i.e. the content of the WebResourceInterface instance.
      *
      * @return string
      */
     public function getDocumentCharacterSet(): string;
-
-    /**
-     * Returns the character set found in the ResponseInterface instance.
-     *
-     * @return string|null
-     */
-    public function getResponseCharacterSet(): ?string;
 }
